requirements.txt
asgiref==3.2.3
django>=3.0.3
django-cors-headers==3.2.0
djangorestframework==3.11.0
gcloud==0.17.0
googleapis-common-protos==1.6.0
httplib2==0.15.0
jws==0.1.3
mysqlclient==1.4.6
oauth2client==3.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycryptodome>=3.6.6
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2019.3
requests>=2.20.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0.1
django-heroku
gunicorn
